backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aiocache==0.12.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# LLM Integrations
openai==1.58.1
anthropic==0.40.0
google-generativeai==0.8.3
langchain==0.0.340
tiktoken==0.5.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
pyarrow==14.0.1

# File handling
python-magic==0.4.27
openpyxl==3.1.2
xlrd==2.0.1
aiofiles==23.2.1
chardet==5.2.0
jsonpath-ng==1.6.1

# Multi-modal Processing
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
matplotlib==3.8.2
python-docx==1.1.0
python-pptx==0.6.23

# Async & Background Tasks
celery==5.3.4
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0